requirements.txt
discord.py
pytube
pynacl
youtube-dl
yt_dlp
python-dotenv
PyYAML
pymongo